Avoid Hodgepodge conflict
